Fixes: rename NetworkRequester typ eto be more clear + split discovery into own file
