+++ sdk/rust/nym-sdk/src/error.rs
@@ -145,7 +145,7 @@ pub enum Error {
     #[error("no available gateway in the requested countries")]
     NoGatewayInCountries,
 
-    #[error("no countries specified; use NetworkRequester::any() to accept any country")]
+    #[error("no countries specified; use NetworkRequesterSelector::any() to accept any country")]
     NoCountriesSpecified,
 
     #[error("invalid network requester address: {0}")]
